Use the SSL certificate when calling S3#download
Sometimes aws-sdk users will want to set

Aws.config[:ssl_ca_bundle] = '/path/to/self-signed-cert.pem'

This automatically uses the certificate on all aws-sdk operations.
However, this certificate is not used in S3#download, since for
downloading open-uri is used and not aws-sdk. So we forward the SSL
certificate from Aws.config to open-uri in case it's set.