Deprecate moving an IO which isn't movable
If we setup moving plugin and IO comes along for which the storage
doesn't support moving, the current behaviour was to "imitate" moving by
uploading and deleting. However, this feature is redundant as it can be
achieved with the delete_uploaded plugin, at least for the raw files. As
for UploadedFiles, we will probably add a plugin which supports deleting
them, or extend delete_uploaded with this, but it should definitely not
be part of the moving plugin.