Generalize Attacher#promote by allowing phase
I initially thought that promoting will only ever be used for cached
files. However, Attacher#promote can actually be used to reupload any
current attachment, and securely swap it with the current attachment. In
this case we need to allow users to pass a custom phase, base when phase
is "store" the file would get reprocessed, likely also causing an error.

One use case for custom promoting is if the location of our attachments
are generated based on the record's slug. When the slug changes, we
might want to change the location of the attachment, ideally in a
background job. Now we can simply reuse the existing job:

  class User < Sequel::Model
    def after_commit
      if (previous_changes || {}).key?(:slug) && image_attacher.stored?
        data = Shrine::Attacher.dump(image_attacher).merge("phase" => "regenerate")
        PromoteJob.perform_async(data)
      end
    end
  end

We also change the suggested naming from "UploadJob" to "PromoteJob" to
generalize its meaning. Besides, this job is also often processing the
files as well, so "UploadJob" isn't an accurate name anyway.