Change direct_upload endpoint to /:storage/upload
Having this route as /:storage/:name was a wrong call. I thought that
it's useful to have as much context as possible inside the uploader.
However, having the name of the attachment is actually not worth
anything without the record, which isn't present in direct uploads.
Furthermore, branching logic depending on the name of the attachment is
a code smell, instead it should be separated into different uploaders.