Add ability to generate fake presigns for testing
This change makes it possible for direct_upload endpoint to generate
"fake" presigns for storages which don't support them. The generated
presign points to the direct upload route, which is now enabled when
storage doesn't support presigns but presign is set. When a file is
direct uploaded this way, we allow setting the location via the "key"
query parameter, to resemble S3.

This fake presigns are very useful for testing, because they allow you
to simply swap out the S3 storage for FileSystem, which keeps your tests
faster and doesn't require any conditionals around presigns.