Accept presign options via :presign_options
Instead of making the :presign option play a double role of choosing
presigns and setting presign options, it's much more explicit to have a
separate :presign_options option for the latter case. While we're at it,
we allow :presign_options to be assigned a static hash of options in
addition to a block.

We also deprecate passing options block to :presign.