Change parallelize plugin to use threads directly
Requiring a dependency just for a simple functionality of thread pools
is an overkill. So we swap ruby-thread's implementation with our own
simple thread pool, effectively removing this dependency while
maintaining the same behaviour.

An initial idea was to switch from ruby-thread to concurrent-ruby, since
concurrent-ruby is now a dependency of a lot of popular projects, like
Rails, Sidekiq and SuckerPunch. However, I couldn't make
concurrent-ruby's thread pool to simply propagate errors. Furthermore,
even if we're ok with silencing errors, concurrent-ruby tries to log the
error, but failing because the logger isn't properly set. The only way
to get around this was to require the whole concurrent-ruby gem, which
is a lot of code if I only want the thread pool. So bare threads it is.