Return Content-Length header in download_endpoint
Returning the Content-Length is really useful e.g. for downloading
files, where the browser can know how much bytes are remaining. It's of
course useful for other things, it's common curtesy to return
Content-Length.

We allow Storage#stream to include in the second yielded argument the
total content length. This should generally be available when file is
streamed from an URL, because Content-Length is read before the content
is starting to be read.