Don't error validations when dimensions are absent
If a user uploads something that isn't an image which you cannot extract
dimensions from, dimensions validations would error because dimensions
would be nil and cause inequality operators to raise an exception. This
is not the behaviour that we want, we don't want that a user can make
validations cause an exception in the application.

Instead in the absence of dimensions we simply don't validate them,
since dimensions validations should only be used for validating
dimensions that exist. The developer should always add an additional
MIME type validation in this case anyway, which will catch non-images.