Rename restore_cached to restore_cached_data
The "restore_cached" name indicates that somehow the cached file itself
is somehow restored. The "restore_cached_data" is accurate, because this
is exactly what the plugin is doing. I was thinking also about
"restore_cached_metadata", but I think that would indicate we're only
restoring metadata if possible, but in this plugin we're ensuring the
cached file actually exists.