Move before_* and after_* hooks around around_*
Current behaviour was that before_* and after_* hooks happen inside the
around_* hook, which means the around_* hook was wrapping before_* and
around_*. However, since the logging plugin uses an around_* hook, this
meant that any user-defined before_* and after_* hooks in the uploader
would trigger *inside* logging. And this is what we wanted to avoid.

So we change the order of hooks to be before_* => around_* => after_*.
This makes more sense anyway, because we kind of expect that what we
define in before_* and after_* hooks will indeed happen before/after,
regardless of the around_* hook. We want the around_* hooks to behave as
an "inner" hook.