Don't use #mime_type for extracting MIME type
This was the behaviour when Shrine#extract_metadata was receiving
Shrine::UploadedFiles. The determine_mime_type plugin was actually using
mime_type as an identifier that it's a Shrine::UploadedFile, and that it
shouldn't extract metadata.

However, now we're simply copying over metadata hash from one
UploadedFile to another, Shrine#extract_metadata never even gets called,
so we can safely remove this.