Bring back promoting triggering callbacks
In 1.4.0 we switched promoting to updating through datasets/relations,
to eliminate a tiny possibility of a race condition, where a new
attachment can be replaced by an old one. However, this means that
callbacks aren't triggered anymore for promoting, which is a breaking
change as users were relying on this. This is wanted behaviour, we just
accidentally broke it in 1.4.0.

We switch back to updating through a model instance, and also tests
around callbacks. For ActiveRecord this brings back the chance of a race
condition, but for Sequel we eliminated this possibility. For
ActiveRecord optimistic locking can be used, and we add rescuing the
errors to the plugin.