Fixed some issues around promoting
This commit reverts the behaviour of promoting as it was in 1.4.0.
Version 1.4.1 introduced some issues: for ActiveRecord on optimistic
locking the attachment would be swapped anyway even if it updated in the
meanwhile, and for Sequel it modified the record instance with an
instance_filter for the cached attachment, which means the same instance
can't be used anymore for any updating, which is more of an issue when
not using backgrounding.

In general one change between 1.4.0 is that Sequel plugin now rescues
only Sequel::Error's which were supposed to be
Sequel::NoExistingObject's, not all Sequel::Errors.

I don't think we will do any further changes around optimistic locking
in version 2.0.0, because there are too many edge cases, it's better
that people handle this themselves.