Stop testing Rubinius
Rubinius has mutable versions, meaning new code can be republished on
the same version, which means behaviour of version x.y today might not
be the same as tomorrow. Sometimes the compilation of Rubinius even
fails.

Since I can't see an advantage of using Rubinius over JRuby, I decided
to stop supporting it. This removes some disabled tests because a
3rd-party library didn't support Rubinius. It also removes some cases
where we had to explicitly pass method arguments to `super` because it
didn't work correctly with keyword arguments.