Remove "fake moving" from moving plugin
The moving plugin was previously playing a double role: if the storage
supports moving, the file is moved, and if it doesn't, the file is
normally uploaded but it's also deleted afterwards. This was adding
complication to this plugin, and not really making it single purpose.
And it's possible now to delete files after uploading with delete_raw
and delete_promoted plugins.

We also make `:storages` optional, and by default it's applied to all
storages.

Finally, we update the documentation to remove unnecessary complications
in the description and make it simple as it is now.