Make default generated unique location shorter
We just use the default length of `SecureRandom.hex`, which is 32, where
before it was 60. This makes it short enough so that we can use it in
pretty_location plugin, but still long enough that we can be sure there
won't be any clashes (it's still 3 * 10^38 combinations).

We also update the direct_upload plugin to use whatever unique generator
the uploader itself uses, in case the user decides to change it.