Improve README
We remove "users" and "avatars", because these are old terms.  People
should never have a "User" in the system, only an "Account" (from
RubyTapas #346), and "avatar" should be named "profile image". But I
think instead it's simpler to use "photos" and "images". We also try to
change types of files in examples, so that users don't get the idea that
Shrine is only for images.