Add back extracting filenames from Tempfiles
This is a backwards-incompatible change which we didn't (couldn't)
deprecate before, and it's not justified enough. The argument was that
Tempfile's filename doesn't have any semantic value, but for example it
could be useful to know where the file came from (by reading Tempfile's
basename), and what was the original extension. People could have relied
on the original filename to get the extension (if they don't know about
`UploadedFile#extension`).