Fix file command in determine_mime_type plugin
The `file` utility's option parser is greedy and interprets every
argument that begins with a hyphen as an option. This means pathnames
that start with a hyphen cause the command to fail, either by invalid
options or never passing it required input. For example,
`file -image.jpg` looks like combined options and is parsed as `-i`,
`-m`, `-a` (invalid), etc.

To prevent this, `--` is appended to the command to signify the end of
options and allows subsequent arguments to be excluded from option
parsing.