Update generalized gems' description in gemspec
We change this in the README, to hint that Shrine does not only have
functionality for uploading files. One idea was to use the term "file
management", but that sounds like it manages existing files on the
filesystem, instead of files uploaded by users using any storage.
Anyway, now we also update this description in the gemspec.