Remove unused Roda plugin in direct_upload
The json_parser Roda plugin parses JSON request body and merges the hash
into `request.params`. We added this plugin when we added the "key"
query parameter in the "/upload" POST route. However, the request body
for the upload route will also contain the uploaded file, it must be in
the URL-encoded format by HTTP specification (multipart body cannot be
with JSON). Therefore, this plugin isn't used, so we remove it.