Use @name variable in Attachment#initialize
This removes any visual ambiguity when reading the code. This is because
Attachment#name already exists as a method inherited from Module#name,
and I would rather that we refer to instance variable directly then to a
local variable. This means that the same code can now go to #initialized
and #included hook.