Allow non-Sequel attachments in sequel plugin
We may want to use Shrine in an application where some models are Sequel
models, and others are POROs or maybe some form models. In that case
it's useful to be able to load the sequel plugin globally, but still
generate attachments for other types of models. Technically this was
still possible before, but the model would get Sequel methods defined.
Instead we can detect if we're including the module into a Sequel model,
and then dynamically define Sequel-specific methods to that module,
otherwise we keep it blank.