Support Mongoid in backgrounding plugin
Mongoid records return a BSON::ObjectId when you ask them for an ID,
which isn't serializable to JSON unless you monkey-patch it to be. So we
call `#to_s` to convert it to a string. This should still work for other
ORMs, because I'm not aware of an ORM without automatic typecasting.