Warn when not using determine_mime_type plugin
The "mime_type" metadata field is by default set from "Content-Type"
request header. However, this means that "mime_type" is not guaranteed
to hold the actual MIME type of the uploaded file, because
"Content-Type" can be set by the user, and by default the browser sets
it solely based on file's extension.

This is a security issue, because the attacker can upload a harmful file
with a .png extension, and the "mime_type" will be set to "image/png".
By "harmful" I mean a file that can compromise the server (e.g.  through
ImageMagick). Users are already warned about it in the README, and told
they should probably use determine_mime_type plugin. But now we also
print a warning, so that users don't miss it.

If users don't want to use any external tools (UNIX `file` command) or
gems (mimemagic, filemagic), they can opt-out from determining MIME type
from file content by setting the `:analyzer` to `:default` when loading
the determine_mime_type plugin.