Make only one HTTP request in download_endpoint
What we do in download_endpoint is turn #stream method on the storage
into an Enumerator. We then first call #peek on the Enumerator just to
get the Content-Length, and then we proceed downloading into response
body.

However, Enumerator#each will start executing the #stream method again
from the beginning, regardless of the fact that we already started
executing it with #peek. This results into two HTTP requests: first one
just downloads the first chunk (and then we just leave it hanging?!),
and the second one downloads the whole file.

We fix this by continuing iterating over the enumerator that already
started the method. Enumerator#peek doesn't move the cursor, so the
Enumerator#next iteration properly start from the first chunk, which was
cached by #peek (we also have Rack::Lint which would raise an error if
length of body didn't match Content-Length).

Note that we don't have to rescue StopIteration raised by `stream.next`
inside the Enumerator declaration, because `loop {}` happens to rescue
StopIteration errors (I don't know why, but it's handy).