Don't use Roda streaming in download_endpoint
Roda's streaming plugin was designed for when you have custom streaming
logic. However, all we have is an Enumerable over which we iterate. It
turns out that Rack's specification states the response body exactly
needs to be an object which responds to #each. Usually people return an
array, but we can also simply return our Enumerator.

I compared this locally to and implementation which just downloads the
whole file and returns it via #send_file. I used a video which I
uploaded to S3, and used Puma. The streaming version was much faster; I
could start watching the video already after a few seconds (even though
it hasn't been downloaded yet). With the #send_file version I had to
first wait for the whole video to be downloaded (the "play" button was
disabled during that time). So the streaming really works!

Note that this implementation is functionally exactly the same as what
the streaming plugin provided us by default, it only differs if we pass
in a custom scheduler (e.g. EventMachine). But since streaming already
works for this version, I don't see what benefit would a custom
scheduler give to the download_endpoint plugin.