Make Attacher#cache!, #store!, and #delete! public
This is handy when you want to customize the upload process, and you
want to cache/store/delete with context forwarded to the uploader.

One change we make is that we don't require `:phase` anymore; it was
added internally so that we don't forget to set a phase when deleting in
other plugins, but a phase is not required. Another change that we make
is the ability to pass in any kind of additional context, not just the
`:phase`.