Add Attacher#_delete which backgrounds #delete!
Now that Attacher#delete! was made public, users should be able to use
it synchronously even if the backgrounding plugin is loaded. The
backgrounding plugin now overrides Attacher#_delete instead, and makes
it spawn a background job. This makes it symmetrical to
Attacher#_promote.

This presents a backwards incompatibility for anyone who was using
Attacher#delete! in a plugin, relying that it will spawn a background
job instead of doing a synchronous delete.