Don't cache dynamic storages
I thought that the block will always be deterministic, that the storage
will always be determined solely based on the MatchData object. So I
figured that I might as well cache the instantiated storage, to save on
allocations.

However, I encountered a real-world use case which required dynamically
instantiating storages with definitions stored in the database. This
means that the same storage key can hold different storage depending on
what is in the database at that moment. Since we were caching the result
of the block, storage definition database updates wouldn't get picked
up.

So we remove the caching in order to make the storages truly dynamic,
with the tradeoff of a small performance penalty. I was thinking of
adding a `:cache` option which brings back caching, but I don't think
it's worthwhile.