Deprecate migration_helpers plugin
The idea of this plugin was very fuzzy in the first place. It was just a
collection of helper methods, which weren't useful and numerous enough
to deserve their own plugin. Also, while helping people I've realized
that `Attacher#cached?` and `Attacher#stored?` are so commonly needed to
customize attaching, that I've moved then to base functionality.
`Attacher#update_stored` should be just replaced by its implementation;
even though it saves code-writing, it's much better to write more
explicit code.