Allow combining MIME type analyzers
Prior to this change you couldn't combine other analyzers, because the
lambda specified as a custom analyzer isn't executed in the context of
the Shrine instance. This change calls the custom lambda with the second
argument (optional) which is a hash of existing analyzers.

  plugin :determine_mime_type, analyzer: ->(io, analyzers) do
    analyzers[:mimemagic].call(io) || analyzers[:file].call(io)
  end

This requires that we rewind the IO inside each analyzer, and that we
require dependencies dynamically.