Remove Shrine::Confirm functionality
It isn't necessary that clearing the storage needs to be confirmed;
deleting all records in the database doesn't require confirmation, and
people are just fine with it. I don't like having Shrine::Confirm in the
base functionality, because it's taking up space and memory, and it is
of very little importance.

This is technically a breaking change that we didn't deprecate, but I
think there is a very low chance that `#clear!` is in production code;
it's more likely to be used manually in the console or in scripts.