Change how upload options and metadata are passed
Passing upload options via {"s3" => {...}} was always awkward. I did it
to preserve backwards compatibility with existing metadata hash. The
real way that upload options should always have been passed is as
top-level options, and the metadata hash would just be a special key.

Now the storage #upload and #move signatures should look like this:

  def upload(io, id, shrine_metadata: {}, **upload_options)

This is a breaking change, but I didn't find any way to deprecate it
before; I just now found out about Method#parameters which would
probably make it possible. But all the open-source storages are locked
to version "~> 1.0" (and they will now be updated to "~> 2.0"), so this
only brings backwards incompatibility to internal storages.