Use presigned_url or public_url in CDN-ed URLs
When and S3 object is private, it isn't accessible via a public URL
(without the additional query parameters). That means the CDN-ed URL
didn't work as well, because it doesn't include any query parameters,
regardless of whether the `:public` option is given or not.

We instead make the CDN-ed URL generated from the original URL, keeping
any query parameters. We handle an unidentified edge case where the URL
is returned as path-style (even though :forced_path_style isn't set)
making the CDN-ed URL not work, in a way that we simply subtract the
bucket-name part from the URI path if it's a path-style URL.