Deprecate setter from cached_attachment_data
I previously encouraged folks to use the cached_attachment_data plugin
like this:

  <%= form_for @user do |f| %>
    <%= f.hidden_field :cached_avatar_data %>
    <%= f.field_field :avatar %>
  <% end %>

The problem arises from the fact that Shrine is dependent on the order
that these parameters will be assigned to the model. If someone happens
to use strong_params in the following way:

  params.require(:user).permit(:avatar, :cached_avatar_data)

this will reverse the order of parameters. One side effect is that the
user cannot reupload a new file in case of validation errors,
because the cached (invalid) file will always overwrite the new file
(unless the page is refreshed and the invalid file is gone).

It would be very brittle to require users to be mindful about the order of
parameters, so we deprecate the setter and recommend the following
instead:

  <%= form_for @user do |f| %>
    <%= f.hidden_field :avatar, value: @user.cached_avatar_data %>
    <%= f.field_field :avatar %>
  <% end %>

Since the parameters are now called the same, whatever is in the file
field will always override the hidden field, before it even reaches the
controller.

Fixes #65