Return cached_attachment_data only if changed
This value is only here for setting the hidden field, whose sole purpose
is to retain the cached file in case of validation errors. This means
that the hidden field should only be populated in that case, the cached
data shouldn't be sent if it's already persisted, which can happen with
backgrounding:

If the user would revisit the form after the background job has been
kicked off but before it's finished, the hidden field would be populated
with the cached attachment. If the job would happen to finish in the
meanwhile and the user submits the form, the new stored file would get
overwritten with the old cached file again. This shouldn't generally
cause any errors, the only side effect would be that the background job
is repeated, but I would rather not let that happen just to be safe. And
since this method's purpose is for retaining cached files across
validation errors, I see this as the correct behaviour.