Don't check for method arity in #_enforce_io
Rack::Test::UploadedFile is defined with the following:

  def method_missing(method_name, *args, &block)
    @tempfile.__send__(method_name, *args, &block)
  end

  def respond_to?(method_name, include_private = false)
    @tempfile.respond_to?(method_name, include_private) || super
  end

This means that calling `respond_to?(:read)` will return true, but
calling `method(:read)` will raise a NoMethodError, because the method
\#read technically isn't defined on Rack::Test::UploadedFile.

We don't need to be so precise and check the arity, people will figure
out that `#read` needs to accept arguments. We also remove arguments
from the error message, because it currently says that #read should be
defined with `read(length, outbuf)`, while in reality it should be
defined with `read(length = nil, outbuf)`, so the former is misleading.

Fixes #67