Execute the file command always in the same way
Previously we were branching the way we execute the file command,
depending on whether the IO was a file or not. However, this leads to
inconsistent behaviour; the MIME type could be determined differently on
the original file and on an UploadedFile with restore_cached_data. We
don't want that, we want that the file command always returns the same
result, regardless of type of the given IO (that's why we have this
abstraction in the first place).