Increase magic header and make it configurable
I blindly copy-pasted the length of 1KB from CarrierWave, but I never
really tried to find out how much does the file command or
ruby-filemagic gem actually need to determine MIME type of every file.

So I ran the following in the command line:

  file = File.open("/path/to/file")
  system "file --mime-type -", in: file
  file.pos #=> 262144

For every file, the file command seemed to always read the first 256KB
of the file. This leads us to the conclusion that this is the correct
number of bytes we should always read (or at least it's a more correct
number).

In case this number is not high enough or is too high, we've made this
number configurable.