Pass UploadedFile when reextracting metadata
We had to recently change the result of Storage#open, which means that
the argument to #extract_metadata also changed. However, we want the
argument to always be the same regardless of the storage, so that users
know they have to extract metadata using the IO interface, instead of
using file methods.

The side effect of this is that now #original_filename is preserved,
where before it could be lost. For #size we need to use the opened size
over the one from metadata.

Another advantage is that the user can choose to reuse any custom
metadata instead of reextracting it (e.g. if it would mean downloading
the whole file).