Don't require :names anymore in versions plugin
Initially the versions plugin didn't have the :names option, but then I
added it because I figured that without it an attacker can DoS the app
by changing the hidden field to have a lot of versions with made-up
names.

However, I think it's a much better solution to forbid generating
versions in the cache phase and assigning them. This way we don't have
any problem. I think shouldn't be limiting, because processing on
caching in general can leave a security leak if not careful, and also
with presigned uploads you cannot even generate versions, which is what
everyone should be doing anyway.

This makes it much nicer to manage versions, because the user doesn't
have to update the :names list every time they add or remove a version.