Remove Storage#read
This method is redundant, because it can be replaced by the following,
which has the same functionality:

io = storage.open(id)
content = io.read
io.close

The more methods the storage needs to implement, the more overwhelming
it is for the storage creators. We should keep things as simple as
possible, and give people the least amount of work.