Don't require attachment module in backgrounding
Some users don't want to include Shrine's attachment module to their
modules in order to avoid pollution, they instead prefer to work with
the Shrine::Attacher directly.

The background plugin previously depended on the "<name>_attacher"
method from the attachment module to get the attacher. However, if we
also pass the Shrine class to the background job, we can instantiate the
Shrine::Attacher directly.

If the uploader is an anonymous class, and in that case doesn't have a
name, we fall back to using "<name>_attacher".