Implement UploadedFile#download in terms of #open
We don't need to require storages to implement #download, if we can
download just using #open.

We still leave the possibility for storages to implement #download if
they want to. For example, S3 storage will use aws-sdk to download the
file, since it has the ability to automatically retry the download in
case it fails.