Soft deprecate :phase over :action in context
The :phase attribute was initially either named like that to mark the
notion of "attachment phase". However, over the time when we added
additional "phases", I realized that this was an inaccurate name from
the beginning, that all of these things are rather "actions" that happen
around attaching (especially deleting).

It's a bit unfortunate to do this rename now, but I think good naming
pays off in the long run since it brings clarity.

So we soft-deprecate :phase over :action, in a way that at the moment we
send both parameters, to keep backwards compatibility. In Shrine 3 when
people will mostly transition to the processing plugin, we will
deprecate referencing this context key, and then remove it in Shrine 4.