Set default file/folder permissions to 0644/0755
Different systems can have different default permissions. Moreover,
upgrading the same system can also change the default permissions.

We should make sure in Shrine that we always use the same file and
directory permissions. A sysadmin confirmed that 0644 and 0755 are good
defaults, and these are also the ones that CarrierWave uses.