Handle aborting on change in backgrounding
We already depend on the #find_record method in backgrounding plugin,
which we can use to check whether the attachment has changed. This check
needs to be done only when using backgrounding anyway, so it's nice to
be able to move it into the appropriate plugin. And this way none of the
ORM plugins need to implement this logic, it's all handled
automatically, yay!

Initially I assigned the reloaded record to the @record instance
variable. However, this would mean that a record instance that we had
outside of the attacher wouldn't be updated, which might be surprising
if people are using backgrounding plugin only for deleting. Here we rely
on ORMs ability to only do partial updates of changed fields (which both
ActiveRecord and Sequel support), otherwise the record might be updated
with old fields and invalidate an updated record during backgrounding.