Fix behaviour of copy plugin
Previously when record is duplicated with new attachment, the original
attachment would be marked for deletion, and would be deleted on
finalization. So we explicitly assign the attachment data to nil
beforehand.

We also make no attachments "copied over" as well.

Lastly, we make more improvements on existing tests and add some missing
tests.