Disable caching presign response in direct_upload
Some browsers like Safari will automatically cache GET and HEAD
responses, which breaks the presign endpoint because external services
require each request to have a unique signature. So we tell the client
not to cache the response by sending "Cache-Control: no-store".

The recommended way so far to avoid this was to add a unique query
parameter in the request (e.g. a timestamp). But with this change it
shouldn't be needed anymore.