Don't change permissions of existing directories
Previously if FileSystem storage was initialized with an existing
directory, the storage would attempt to change its permissions. However,
this is not always possible; for example, instantiating the storage with
`Dir.tmpdir` would in this case raise an `Errno::EPERM` exception,
because the operating system doesn't allow changing the permissions of
the temporary directory if you don't have root access.

The correct behaviour should be that we only apply the directory
permissions only for newly created directories, and don't change
permission of existing directories. This way we don't have the risk of
modifying the permissions that were explicitly set outside of Shrine.

Fixes #80