Don't fail in FileSystem if file is missing
It is common that the user clears a storage manually, and in this case
deleting existing records with attachment would fail, because the
storage would raise an error when attempting to delete files that don't
exist.

Since I cannot think of a scenario where "file doesn't exist" error
could be caused by an anomaly in the system (in which case hiding it
would hurt introspectability), I think we can safely ignore this error,
and make Storage#delete a no-op in that case.

We also update the linter to encourage all other storages to adopt this
behaviour as well.

Fixes #87