Remove deleting Tempfiles returned by Storage#open
Before Storage#open returned a Tempfile object for most storages. Even
though the GC is supposed to delete closed Tempfile eventually, in
practice that hasn't been working reliably, especially for JRuby. The
files would only get deleted when the process ends, but not on garbage
collection. So to work around that Shrine additionally deleted the
Tempfiles after closing.

However, now all major storages return Down::ChunkedIO objects, which
automatically delete the underlying Tempfile when `#close` is called.
Therefore this workaround isn't needed anymore.

If there are some storages which still return a Tempfile, they should
consider using Down::ChunkedIO or a similar kind of object. Ultimately
this bug of Tempfiles not being deleted on GC should be solved in the
Ruby implementation itself.