Don't use Storage#stream in download_endpoint
Storage#stream was in the past used for streaming the uploaded file in
chunks. In the meanwhile this concept has been replaced by
Down::ChunkedIO, which is an IO object that automatically streams the
uploaded file as it is read. This enabled us to merge Storage#stream
into Storage#open.

Shrine still used Storage#stream for performance backwards
compatibility, but it's already been many months since the major
storages adopted Down::ChunkedIO, so now we can just rely on
Storage#open.