Don't run validations in copy plugin
Currently we're assigning the copied file using `Attacher#set`, which
runs file validations, and those validations fail when using versions.

When we're copying attachment from another record, that attachment is
expected to have already passed file validations. Therefore we can skip
file validations when assigning a copied file.

An idea was to add a `:validate` keyword argument to `Attacher#set` and
maybe also `Attacher#assign`, which will determine whether the file
validations are run. But that's a more complex change, and could be
backwards incompatible. Since nobody request this option yet, it's
better to do the simpler change.