Allow dynamic options in default_url_options
There are cases where we want to pass the URL options based on
UploadedFile's metadata, but we want to make them the default URL
options. An example would be setting a dynamic "Content-Disposition"
header based on the filename.

This commit extends the default_url_options plugin with the ability to
accept a block, which is called with the UploadedFile object and any
URL options that were passed to `UploadedFile#url`.

  plugin :default_url_options, store: ->(io, **options)
    if options.key?(:download)
      {response_content_disposition: "attachment; filename=#{io.original_filename.inspect}"}
    end
  end

Addresses #91