Handle non-ASCII chars in filename on S3 upload
When we're uploading a file to S3, we're also including the parameter
for the "Content-Disposition" header. However, if the extracted filename
contains any non-ASCII characters, the aws-sdk gem will fail to
calculate the correct signature for upload, and will raise a
SignatureDoesNotMatch error.

https://github.com/aws/aws-sdk-ruby/issues/1230

In order to work around it in Shrine, we remove any non-ASCII characters
by URL-encoding the filename on upload, since some browsers like Chrome
will automatically URL-decode the "Content-Disposition" header on
download.

Even though #url and #presign don't fail with non-ASCII characters, they
still doesn't produce the correct filename when downloaded, so we fix
that as well.

Fixes #93