Support JSON Postgres columns with ActiveRecord
Unlike Sequel, which accepts values as JSON strings on a JSON/JSONB
Postgres column, ActiveRecord requires the value to be a Hash. If it
receives a JSON string (which Shrine does by default), it will treat it
as a single value of JSON string type.

So we override Attacher#write to convert JSON strings into hashes if the
data column is a JSON or JSONB Postgres column.

Fixes #94