Make remove_invalid bring back previous attachment
Previously if the assigned file was invalid, it would get deleted and
`nil` will get assigned instead. However, that means that if there was a
previous attachment, and the user wants to display it in case of
validation errors, they cannot even with the remove_invalid plugin.

I think that this isn't correct behaviour. The goal of this plugin
should be to bring the record in the state as it was before the
assignment of the invalid file. This means that if there was a previous
attached file, it should be assigned back. This is what this change
accomplishes.

Fixes #96