Perform JSON conversion in #_set instead of #write
Prior to version 2.4.0, the JSON conversion of data was performed in
Attacher#_set. In Shrine 2.4.0 we accidentally moved this to
Attacher#write, introducing a regression for anyone who was using
Attacher#write directly and passing it a JSON string (resulting in a
double-encoded string).

We ultimately want to move #convert_before_write back into
Attacher#write, so that it's symmetrical with Attacher#read, but that
will have to wait for the next major release.

Fixes #97