Allow extracting multiple custom metadata at once
Some tools extract multiple metadata values at once (streamio-ffmpeg,
MiniMagick etc). Previously it was possible to only assign a single
metadata value per block, which means that in order to extract multiple
values, we would have to call the tool multiple times.

This is inefficient, because even though for these tools a single call
already extracts all the metadata that we need, we still have to repeat
it multiple times. And extracting metadata can be an expensive process,
so it's not ideal to have to call it multiple times.

This commit extends the `add_metadata` plugin to add the ability for
extracting multiple metadata values in a single block.

  add_metadata do |io, context|
    movie = FFMPEG::Movie.new(io.path)

    { "duration"   => movie.duration,
      "bitrate"    => movie.bitrate,
      "resolution" => movie.resolution,
      "frame_rate" => movie.frame_rate }
  end

In this case we won't automatically get the reader methods on
UploadedFile, but we can create them via `metadata_method`:

  metadata_method :duration, :bitrate, :resolution, :frame_rate