Allow disabling fallback to original with versions
When versions are processed in a background job, requesting the version
URL via `Attacher#url` before it has been processed will fall back to
the original cached file.

  photo.image #=> <Shrine::UploadedFile>
  photo.image_url(:large) # falls back to `photo.image_url`

This enables automatic graceful degradation for the period when
background job is still processing the original file. However, since the
cached file can be very big, sometimes you'd rather want to just display
some default URLs which indicate to the user that versions are being
processed.

This commit adds a :fallback_to_original option to the versions plugin,
which can be set to false if you want to disable this fallback.

  plugin :versions, fallback_to_original: false

This allows you to set up default URLs for versions which are going to
be processed in a background job.

  plugin :default_url do |context|
    "defaults/#{context[:version]}.png"
  end